Crate rza1

Peripheral access API for RZ/A1H microcontrollers (generated using svd2rust v0.14.0)

You can find an overview of the API here.

Cargo Features

This crate defines the following Cargo features: adc, bsc, ceu, cpg, disc, dmac, dvdec, ether, flctl, gpio, ieb, inb, intc, irda, jcu, l2c, lin, lvds, mlb, mmc, mtu2, ostm, pfv, pwm, riic, romdec, rscan, rspi, rtc, scif, scim, scux, sdg, spdif, spibsc, ssif, usb20, vdc5, wdt

All features are disabled by default.

Modules

ostm0
OSTM0

Structs

OSTM0
OSTM0

OSTM1
OSTM1

Peripherals
All the peripherals